Bug Fixed in camera controller.
Camera controller had a bug that was only identified after the simple respawn was added.
When checking if theres an object between the player and camera, it worked as follows:
- If raycast hits:
   - if hit environment:
     - if hit distance > desired distance:
          current distance = desired distance
      else:
          current distance = hit distance
else:
    current distance = desired distance.

this left 1 gap, as if the raycast did hit something that wasnt tagged as environment, it wouldnt update the cameras distance at all.
The respawn zones were causing this to happen, meaning if the player had one between them and the camera, they could zoom in and out without issue, but if a wall was hit first, then this zone, the camera wouldnt re adjust for the absence of wall, as it never entered the section comparing the tag, and never left the raycast check.