Updated respawn mechanic
now when the player dies, instead of respawning after x amount of time, they have 2 buttons, either respawn or main menu. no functionality behind main menu right now.
Also instead of getting a ref to the player when the game starts, instead the game manager spawns them in on start. this will mean when the game is loaded from a save file, they can be spawned in at the correct position instead of having to be moved after they spawn, which will cause issues with the character controller.